newsDataUpdate/requirements.txt
# Function dependencies, for example:
# package>=version

Flask==1.0.2
requests==2.21.0
beautifulsoup4==4.4.0
google-cloud-storage==1.18.0